requirements.txt
Flask==3.1.3
flask-login==0.6.3
gunicorn==25.3.0
beautifulsoup4==4.13.4
jsonschema==4.25.1
pytest==8.3.4
python-dotenv==1.1.0
pytest-flask==1.3.0
